# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hono API",
    "version": "1.0.0",
    "description": "API documentation for the Hono project."
  },
  "tags": [
    { "name": "Auth", "description": "User authentication endpoints" },
    { "name": "Books", "description": "Book management endpoints" },
    { "name": "Authors", "description": "Author management endpoints" },
    { "name": "API Keys", "description": "API key management endpoints" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered." },
          "409": { "description": "Email already in use." }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT token returned." },
          "401": { "description": "Invalid email or password." }
        }
      }
    },
    "/books": {
      "get": {
        "tags": ["Books"],
        "summary": "List all books",
        "responses": {
          "200": { "description": "A list of books." }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "Create a book (API key required)",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "minLength": 1 },
                  "description": { "type": "string" },
                  "publishDate": { "type": "string", "format": "date" },
                  "pageCount": { "type": "integer", "minimum": 1 },
                  "authorId": { "type": "string", "format": "uuid" }
                },
                "required": ["title", "authorId"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Book created." },
          "400": { "description": "Author not found." }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Book found." },
          "404": { "description": "Book not found." }
        }
      },
      "put": {
        "tags": ["Books"],
        "summary": "Update book by ID (API key required)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "minLength": 1 },
                  "description": { "type": "string", "nullable": true },
                  "publishDate": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "pageCount": {
                    "type": "integer",
                    "minimum": 1,
                    "nullable": true
                  },
                  "authorId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Book updated." },
          "400": { "description": "Author not found." },
          "404": { "description": "Book not found." }
        }
      },
      "delete": {
        "tags": ["Books"],
        "summary": "Delete book by ID (API key required)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "responses": {
          "204": { "description": "Book deleted." },
          "404": { "description": "Book not found." }
        }
      }
    },
    "/api-keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys for user (JWT required)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "A list of API keys." }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create a new API key (JWT required)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Expiration timestamp for the API key (optional)"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "API key created." }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Delete API key by ID (JWT required)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "API key deleted." }
        }
      }
    },
    "/authors": {
      "get": {
        "tags": ["Authors"],
        "summary": "List all authors",
        "responses": {
          "200": {
            "description": "A list of authors."
          }
        }
      },
      "post": {
        "tags": ["Authors"],
        "summary": "Create an author (API key required)",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "birthday": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created."
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "tags": ["Authors"],
        "summary": "Get author by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Author found." },
          "404": { "description": "Author not found." }
        }
      },
      "put": {
        "tags": ["Authors"],
        "summary": "Update author by ID (API key required)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "birthday": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Author updated." },
          "404": { "description": "Author not found." }
        }
      },
      "delete": {
        "tags": ["Authors"],
        "summary": "Delete author by ID (API key required)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "API key for authentication"
          }
        ],
        "responses": {
          "204": { "description": "Author deleted." },
          "404": { "description": "Author not found." }
        }
      }
    }
  }
}